+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Nonzero trailing bits are allowed when deserializing JSON bytes from base64.
 - Serialization of `google.protobuf.FieldMask` fails if the path cannot be roundtripped through camelCase.
 - `google.protobuf.Duration` uses a different number of trailing zeroes depending on the precision of the input.
+- JSON serialization of `google.protobuf.Timestamp` fails if it is outside the allowed range.
 
 ## [0.3.5] - 2022-01-02